some optimizations
