+++ packages/snfoundry/scripts-ts/helpers/fees.ts
@@ -1,29 +1,7 @@
-import { Account, Contract, Provider, uint256 } from "starknet";
+import { Contract, Provider, uint256 } from "starknet";
 import { Abi } from "starknet";
 import { red, yellow } from "./colorize-log";
-
-export const ETH_TOKEN_ADDRESS =
-    "0x049d36570d4e46f48e99674bd3fcc84644ddd6b96f7c741b1562b82f9e004dc7";
-export const STRK_TOKEN_ADDRESS =
-    "0x04718f5a0fc34cc1af16a1cdee98ffb20c31f5cd61d6ab07201858f4287c938d";
-
-const ETH_TOKEN_ADDRESS_DEVNET = '0x49D36570D4E46F48E99674BD3FCC84644DDD6B96F7C741B1562B82F9E004DC7'
-const STRK_TOKEN_ADDRESS_DEVNET = '0x04718f5a0fc34cc1af16a1cdee98ffb20c31f5cd61d6ab07201858f4287c938d'
-
-const feeTokenOptions = {
-    devnet: [
-        { name: "eth", address: ETH_TOKEN_ADDRESS_DEVNET },
-        { name: "strk", address: STRK_TOKEN_ADDRESS_DEVNET },
-    ],
-    mainnet: [
-        { name: "eth", address: ETH_TOKEN_ADDRESS },
-        { name: "strk", address: STRK_TOKEN_ADDRESS },
-    ],
-    sepolia: [
-        { name: "eth", address: ETH_TOKEN_ADDRESS },
-        { name: "strk", address: STRK_TOKEN_ADDRESS },
-    ]
-};
+import { Network } from "../types";
 
 export const erc20ABI = [
     {
@@ -47,13 +25,13 @@ export const erc20ABI = [
 
 //function to decide preferred token for fee payment
 export async function getTxVersion(
-    deployer: Account,
-    provider: Provider,
+    network: Network,
     feeToken: string,
-    network: string
 ) {
+    const { feeToken: feeTokenOptions, provider, deployer } = network;
+
     //check the specified feeToken
-    const specifiedToken = feeTokenOptions[network].find(
+    const specifiedToken = feeTokenOptions.find(
         (token) => token.name === feeToken
     );
     if (specifiedToken) {
@@ -72,7 +50,7 @@ export async function getTxVersion(
     }
 
     // Check other options
-    for (const token of feeTokenOptions[network]) {
+    for (const token of feeTokenOptions) {
         if (token.name !== feeToken) {
             const balance = await getBalance(
                 deployer.address,
