+++ packages/snfoundry/scripts-ts/helpers/networks.ts
@@ -20,6 +20,9 @@ const deployerDevnet =
   PRIVATE_KEY_DEVNET &&
   new Account(providerDevnet, ACCOUNT_ADDRESS_DEVNET, PRIVATE_KEY_DEVNET, "1");
 
+const ETH_TOKEN_ADDRESS_DEVNET = '0x49D36570D4E46F48E99674BD3FCC84644DDD6B96F7C741B1562B82F9E004DC7'
+const STRK_TOKEN_ADDRESS_DEVNET = '0x04718f5a0fc34cc1af16a1cdee98ffb20c31f5cd61d6ab07201858f4287c938d'
+
 // sepolia
 const providerSepolia =
   process.env.RPC_URL_SEPOLIA &&
@@ -33,6 +36,11 @@ const deployerSepolia =
     process.env.PRIVATE_KEY_SEPOLIA,
     "1"
   );
+
+const ETH_TOKEN_ADDRESS =
+  "0x049d36570d4e46f48e99674bd3fcc84644ddd6b96f7c741b1562b82f9e004dc7";
+const STRK_TOKEN_ADDRESS =
+  "0x04718f5a0fc34cc1af16a1cdee98ffb20c31f5cd61d6ab07201858f4287c938d";
 
 // mainnet
 const providerMainnet =
@@ -48,8 +56,24 @@ const deployerMainnet =
     "1"
   );
 
+
+const feeTokenOptions = {
+  devnet: [
+    { name: "eth", address: ETH_TOKEN_ADDRESS_DEVNET },
+    { name: "strk", address: STRK_TOKEN_ADDRESS_DEVNET },
+  ],
+  mainnet: [
+    { name: "eth", address: ETH_TOKEN_ADDRESS },
+    { name: "strk", address: STRK_TOKEN_ADDRESS },
+  ],
+  sepolia: [
+    { name: "eth", address: ETH_TOKEN_ADDRESS },
+    { name: "strk", address: STRK_TOKEN_ADDRESS },
+  ]
+};
+
 export const networks: Networks = {
-  devnet: { provider: providerDevnet, deployer: deployerDevnet },
-  sepolia: { provider: providerSepolia, deployer: deployerSepolia },
-  mainnet: { provider: providerMainnet, deployer: deployerMainnet },
+  devnet: { provider: providerDevnet, deployer: deployerDevnet, feeToken: feeTokenOptions.devnet },
+  sepolia: { provider: providerSepolia, deployer: deployerSepolia, feeToken: feeTokenOptions.sepolia },
+  mainnet: { provider: providerMainnet, deployer: deployerMainnet, feeToken: feeTokenOptions.mainnet },
 };
